update post counts when deletion happens

+++ app/activitypub/routes.py
@@ -504,8 +504,12 @@ def process_inbox_request(request_json, activitypublog_id, ip_address):
                         user_ap_id = request_json['object']['actor']
                         community_ap_id = request_json['object']['audience'] if 'audience' in request_json['object'] else request_json['actor']
                         to_be_deleted_ap_id = request_json['object']['object']
-                        delete_post_or_comment(user_ap_id, community_ap_id, to_be_deleted_ap_id)
-                        activity_log.result = 'success'
+                        if isinstance(to_be_deleted_ap_id, dict):
+                            activity_log.result = 'failure'
+                            activity_log.exception_message = 'dict instead of string ' + str(to_be_deleted_ap_id)
+                        else:
+                            delete_post_or_comment(user_ap_id, community_ap_id, to_be_deleted_ap_id)
+                            activity_log.result = 'success'
                     elif request_json['object']['type'] == 'Page': # Editing a post
                         post = Post.query.filter_by(ap_id=request_json['object']['id']).first()
                         if post:
@@ -685,12 +689,14 @@ def process_inbox_request(request_json, activitypublog_id, ip_address):
                     post = Post.query.filter_by(ap_id=ap_id).first()
                     if post:
                         post.delete_dependencies()
+                        post.community.post_count -= 1
                         db.session.delete(post)
                     else:
                         reply = PostReply.query.filter_by(ap_id=ap_id).first()
                         if reply:
                             reply.body_html = '<p><em>deleted</em></p>'
                             reply.body = 'deleted'
+                            reply.post.reply_count -= 1
                     db.session.commit()
                     activity_log.result = 'success'
                 elif request_json['type'] == 'Like':  # Upvote
